test: Added additional tests and moved steale test to its own directory
