+++ tests/api/mod.rs
@@ -1,3 +1,4 @@
 mod archive_basic_test;
 mod archive_multihost_test;
 mod archive_multijursidiction_test;
+mod stelae;
